requirements.txt
absl-py==0.3.0
astor==0.7.1
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.6.0
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
idna==2.7
imageio==2.4.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
numpy==1.16.4
olefile==0.45.1
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==5.3.0
protobuf==3.6.0
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.13
requests==2.19.1
resources==0.0.1
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
staty==0.5.0
spectral==0.19
sphinx==2.0.0
tensorboard==1.9.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
toolz==0.9.0
urllib3==1.23
Werkzeug==0.14.1
